fix(core): guard fingerprint blob length check against int overflow
FromBlob validated blob length with 6 + count * 4 in int arithmetic, where count comes from an untrusted (corruptible) cache blob. A count of 0x40000000 wraps the multiply to 0, passes the length guard, and reaches new uint[1073741824], a ~4 GB allocation the runtime rejects with OutOfMemoryException that no caller catches. Compute the check in long arithmetic so an overflowing count is rejected as a cache miss.

+++ src/Spektra.Core/Fingerprint.cs
@@ -29,7 +29,9 @@ public sealed record Fingerprint(byte FramesPerSecond, uint[] Words)
     {
         if (blob.Length < 6 || blob[0] != BlobVersion) return null;
         var count = BinaryPrimitives.ReadInt32LittleEndian(blob.AsSpan(2));
-        if (count < 0 || blob.Length != 6 + count * 4) return null;
+        // (long) so a hostile count near int.MaxValue / 4 cannot overflow the
+        // multiply and slip a huge allocation past the length check.
+        if (count < 0 || blob.Length != 6L + (long)count * 4) return null;
         var words = new uint[count];
         for (var i = 0; i < count; i++)
             words[i] = BinaryPrimitives.ReadUInt32LittleEndian(blob.AsSpan(6 + i * 4));
